winewayland.drv: Handle Wayland surface output positioning.
Track the output(s) a Wayland surface is visible in, to allow us to
update the position of the associated window in the Windows virtual
screen space.

We always place windows at the top-left origin of their main output. The
main output is updated to be the topleft-most output fully contained by
the surface in the fullscreen case, otherwise it's the topleft-most output
entered by the surface.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>