winewayland.drv: Apply the output scale to Wayland surfaces.
Use the scale of the main output of each Wayland surface to always
render at the native resolution on that output. This effectively turns
off compositor-side scaling, giving applications/Wine complete control
over how to scale.

We don't use wl_surface_set_buffer_scale, but instead rely on
wp_viewport to apply the proper scaling, so that we can deal with
fractional scaling. We need this even for integer scales because
wl_surface_set_buffer_scale requires buffer sizes to be integer
multiples of the scale, but since the buffer size corresponds to the
window size, this is not a requirement we can always uphold.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>