winewayland.drv: Explicitly set the Wayland surface size.
Use the viewporter protocol to explicitly set the size of surfaces on
the Wayland compositor side, possibly scaling if needed.

This ensures that regardless of the buffer size we end up committing in
a surface, we will always end up with the expected size, avoiding
protocol errors for states that have strict size constraints (e.g.,
maximized).

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>