winewayland.drv: Avoid visual artifacts during intensive resizes.
We are handling the xdg configure event in the context of a driver
MsgWaitForMultipleObjectsEx invocation, which may have blocked for some
time before the event arrived. Since the last time the message loop was
marked as idle (in terms of window_surface flushing) was before such
wait, if we post the configure message we will continue without allowing
the loop to be considered idle. This may lead Wine core to think that the
app never goes idle (see flush_window_surfaces in win32u), and thus start
flushing at unfortunate times (e.g., in between partial window paints),
causing visual artifacts.

To mitigate this, we post the configure with a small delay to give the
loop some breathing space to be considered idle again.

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>