winewayland.drv: Apply the output scale to Wayland cursor surfaces.
Use the scale of the main output of each Wayland surface to render
cursors at the native resolution on that output.

We perform the scaling with a wp_viewport, but also use
wl_surface_set_buffer_scale for compositors that don't support
wp_vieport for cursors.

Note that when using wl_surface_set_buffer_scale the buffer size is
required to be an integer multiple of the scale, but there are
fundamental issues with libwayland-cursor (see [1]) that block
compositors from strictly enforcing the spec in this case. For now we
can keep using the buffer scale for cursors.

[1] https://gitlab.freedesktop.org/wayland/wayland/-/issues/194

Signed-off-by: Alexandros Frantzis <alexandros.frantzis@collabora.com>